backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
openai==2.14.0
python-dotenv==1.0.1
psycopg2-binary==2.9.11
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<5.0.0
python-multipart==0.0.9
email-validator==2.1.0
requests==2.31.0
